Clear out gratipay/__init__.py
