+++ gratipay/exceptions.py
@@ -52,6 +52,15 @@ class ProblemChangingNumber(Exception):
         return self.msg
 
 
+class NotSane(Exception):
+    """This is used when a sanity check fails.
+
+    A sanity check is when it really seems like the logic shouldn't allow the
+    condition to arise, but you never know.
+
+    """
+
+
 class TooGreedy(Exception): pass
 class NoSelfTipping(Exception): pass
 class NoTippee(Exception): pass
